third_party/blink/web_tests/platform/generic/http/tests/eventsource/eventsource-status-code-states-expected.txt

Test EventSource states for different status codes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 200 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 204 resulted in states CONNECTING, CLOSED, CLOSED
PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 205 resulted in states CONNECTING, CLOSED, CLOSED
PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 202 resulted in states CONNECTING, CLOSED, CLOSED
PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 301 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 302 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 303 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 307 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS arrayCompare(states, tests[count].expectedStates) is true
PASS status code 404 resulted in states CONNECTING, CLOSED, CLOSED
DONE
PASS successfullyParsed is true

TEST COMPLETE
